Remove the run_constrained requirement for jlab.
Fixes https://github.com/conda-forge/ipywidgets-feedstock/issues/65

Since jupyterlab_widgets is now installed by default with ipywidgets, having this run_constrained line in would insist that ipywidgets cannot be used with jlab 2, which is overreaching.

For now, it seems that consensus is that it is better to have the dependency between ipywidgets and jupyterlab_widgets (making things much more convenient for users) than having things more pure from a dependency standpoint.

We may change our mind in the future, though. For example, when jupyterlab_widgets is upgraded to support jlab 4, then blindly installing jupyterlab_widgets may not work for jlab 3, unless we do some work to support both jlab 3 and jlab 4 in the same package.